added support for custom compute offering
